15-85 Split
this definitely does not work. i had a gpt write it and didn't even look at the code before copy/pasting it in here.

basically going to use this as a framework and build out from here. fully expect this to be a huge mistake and waste a bunch of time.